Add the logic to detect available command line options in tools
The goal of this new (internal) feature is to be able to generate the best
command lines depending on what the available tools version provide.

With this we can now support, for example, new ffmpeg versions without breaking
the compatibility with old ones.

This is implemented via the new GetCapabilityAction and a new helper function:
toolbox2.utils.is_option_available()

Internally, it parses the help text/stdout of the tool.

Currently supported tools: ffmpeg, ffprobe